Consolidating product based views to the same file
